+++ src/components/shared/Dialogs/index.ts
@@ -1,0 +1,4 @@
+export { default as ConfirmationDialog } from "./ConfirmationDialog";
+export { default as DraggableDialog } from "./DraggableDialog";
+export { default as NewExperimentDialog } from "./NewExperimentDialog";
+export { default as PipelineNameDialog } from "./PipelineNameDialog";
